refactor(sync): split ProfileDataSyncHandler to respect SwiftLint limits
CI was failing `just format-check` after this branch pushed the handler past
the file-length (1000) and type-body-length (350) thresholds. Rather than
re-baseline the violation, split the handler into focused extension files and
refactor the functions that previously had baselined complexity/body-length
violations so they no longer trigger them in their new file paths (baseline
entries key off file path and don't follow moved code).

Production:
- `ProfileDataSyncHandler.swift` keeps the class shell, properties, init,
  `buildCKRecord`, shared `fetchOrLog`/`batchLogger` helpers.
- `ProfileDataSyncHandler+ApplyRemoteChanges.swift` owns `applyRemoteChanges`,
  `applyBatchSaves`, `applyBatchDeletions`. The former ten-case switches in
  `applyBatchSaves`/`applyBatchDeletions` become dispatch-table lookups
  (`batchUpserters`, `uuidDeleters`), dropping cyclomatic complexity to 1 and
  keeping the functions short. `preExtractedSystemFields` becomes `[] = []`
  (non-optional with empty default), resolving the
  `discouraged_optional_collection` baseline.
- `ProfileDataSyncHandler+BatchUpsert.swift` holds the per-type upsert
  helpers; short closure names (`ck`) are replaced by `$0`/`record`.
- `ProfileDataSyncHandler+RecordLookup.swift` houses `buildBatchRecordLookup`
  (now a straight-line sequence of typed `lookupX` helpers over a shared
  `lookupAndRemove`), `recordToSave` (finder array instead of if-let chain),
  and the per-type `fetch*` helpers.
- `ProfileDataSyncHandler+QueueAndDelete.swift` holds `queueAllExistingRecords`,
  `queueUnsyncedRecords`, `deleteLocalData`. Short `r` names become `record`.
- `ProfileDataSyncHandler+SystemFields.swift` consolidates the former
  `update*`/`clear*` pairs into `setEncodedSystemFields(_:data:recordType:context:)`
  (and the Instrument equivalent) dispatched through a single
  `systemFieldSetters` table, and splits `handleSentRecordZoneChanges` into
  short helpers.

Tests:
- Split the ProfileDataSyncHandler tests across three files
  (`ProfileDataSyncHandlerTests` for remote-changes and buildCKRecord,
  `ProfileDataSyncHandlerQueueTests` for queue/delete/clear,
  `ProfileDataSyncHandlerLookupTests` for lookup/recordToSave) so no struct
  body exceeds 350 lines. The shared `makeHandler` fixture moves to
  `MoolahTests/Support/ProfileDataSyncHandlerTestSupport.swift`.

No behaviour change: `just test` and `just format-check` both pass locally,
and `.swiftlint-baseline.yml` is untouched.
