+++ Backends/CloudKit/Sync/ProfileDataSyncHandler+QueueAndDelete.swift
@@ -1,0 +1,147 @@
+@preconcurrency import CloudKit
+import Foundation
+import OSLog
+import SwiftData
+import os
+
+extension ProfileDataSyncHandler {
+  // MARK: - Queue All Existing Records
+
+  /// Scans all record types in the local store and returns their CKRecord.IDs.
+  /// Called on first start when there's no saved sync state.
+  /// Returns record IDs in dependency order for the coordinator to queue.
+  func queueAllExistingRecords() -> [CKRecord.ID] {
+    let signpostID = OSSignpostID(log: Signposts.sync)
+    os_signpost(
+      .begin, log: Signposts.sync, name: "queueAllExistingRecords", signpostID: signpostID)
+    defer {
+      os_signpost(
+        .end, log: Signposts.sync, name: "queueAllExistingRecords", signpostID: signpostID)
+    }
+
+    var recordIDs: [CKRecord.ID] = []
+
+    // Queue in dependency order:
+    // 1. Instruments (no dependencies)
+    // 2. Categories (no dependencies)
+    // 3. Accounts (no dependencies)
+    // 4. Earmarks (reference instruments)
+    // 5. Budget items (reference earmarks + categories + instruments)
+    // 6. Investment values (reference accounts + instruments)
+    // 7. Transactions (header only)
+    // 8. Transaction legs (reference transactions, accounts, instruments)
+    // 9. CSV import profiles (reference accounts)
+    // 10. Import rules (optionally reference accounts via accountScope)
+    collectAllStringIDs(InstrumentRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(CategoryRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(AccountRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(EarmarkRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(EarmarkBudgetItemRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(InvestmentValueRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(TransactionRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(TransactionLegRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(CSVImportProfileRecord.self, into: &recordIDs) { $0.id }
+    collectAllUUIDs(ImportRuleRecord.self, into: &recordIDs) { $0.id }
+
+    if !recordIDs.isEmpty {
+      logger.info("Collected \(recordIDs.count) existing records for upload")
+    }
+    return recordIDs
+  }
+
+  /// Scans all record types and returns CKRecord.IDs for records that have never been
+  /// successfully sent to CloudKit (i.e. `encodedSystemFields == nil`). Used on startup
+  /// to backfill uploads for profiles whose data landed via migration or any other path
+  /// that bypassed the repository `onRecordChanged` hooks.
+  func queueUnsyncedRecords() -> [CKRecord.ID] {
+    let signpostID = OSSignpostID(log: Signposts.sync)
+    os_signpost(
+      .begin, log: Signposts.sync, name: "queueUnsyncedRecords", signpostID: signpostID)
+    defer {
+      os_signpost(
+        .end, log: Signposts.sync, name: "queueUnsyncedRecords", signpostID: signpostID)
+    }
+
+    var recordIDs: [CKRecord.ID] = []
+    // Same dependency order as queueAllExistingRecords.
+    collectUnsynced(InstrumentRecord.self, into: &recordIDs) { $0.id }
+    collectUnsynced(CategoryRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(AccountRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(EarmarkRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(EarmarkBudgetItemRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(InvestmentValueRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(TransactionRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(TransactionLegRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(CSVImportProfileRecord.self, into: &recordIDs) { $0.id.uuidString }
+    collectUnsynced(ImportRuleRecord.self, into: &recordIDs) { $0.id.uuidString }
+
+    if !recordIDs.isEmpty {
+      logger.info("Collected \(recordIDs.count) unsynced records for upload")
+    }
+    return recordIDs
+  }
+
+  // MARK: - Local Data Deletion
+
+  /// Deletes all local records for this profile's zone.
+  /// Returns the set of all record type strings (for notification).
+  func deleteLocalData() -> Set<String> {
+    let context = ModelContext(modelContainer)
+
+    func deleteAll<T: PersistentModel>(_ type: T.Type) {
+      for record in Self.fetchOrLog(FetchDescriptor<T>(), context: context) {
+        context.delete(record)
+      }
+    }
+
+    deleteAll(InstrumentRecord.self)
+    deleteAll(AccountRecord.self)
+    deleteAll(TransactionRecord.self)
+    deleteAll(TransactionLegRecord.self)
+    deleteAll(CategoryRecord.self)
+    deleteAll(EarmarkRecord.self)
+    deleteAll(EarmarkBudgetItemRecord.self)
+    deleteAll(InvestmentValueRecord.self)
+    deleteAll(CSVImportProfileRecord.self)
+    deleteAll(ImportRuleRecord.self)
+
+    do {
+      try context.save()
+      logger.info("Deleted all local data for profile \(self.profileId)")
+      return Set(RecordTypeRegistry.allTypes.keys)
+    } catch {
+      logger.error("Failed to delete local data: \(error)")
+      return []
+    }
+  }
+
+  // MARK: - Private Helpers
+
+  private func collectAllUUIDs<T: PersistentModel>(
+    _ type: T.Type, into recordIDs: inout [CKRecord.ID], extract: (T) -> UUID
+  ) {
+    let context = ModelContext(modelContainer)
+    for record in Self.fetchOrLog(FetchDescriptor<T>(), context: context) {
+      recordIDs.append(CKRecord.ID(recordName: extract(record).uuidString, zoneID: zoneID))
+    }
+  }
+
+  private func collectAllStringIDs<T: PersistentModel>(
+    _ type: T.Type, into recordIDs: inout [CKRecord.ID], extract: (T) -> String
+  ) {
+    let context = ModelContext(modelContainer)
+    for record in Self.fetchOrLog(FetchDescriptor<T>(), context: context) {
+      recordIDs.append(CKRecord.ID(recordName: extract(record), zoneID: zoneID))
+    }
+  }
+
+  private func collectUnsynced<T: PersistentModel & SystemFieldsCacheable>(
+    _ type: T.Type, into recordIDs: inout [CKRecord.ID], extract: (T) -> String
+  ) {
+    let context = ModelContext(modelContainer)
+    for record in Self.fetchOrLog(FetchDescriptor<T>(), context: context)
+    where record.encodedSystemFields == nil {
+      recordIDs.append(CKRecord.ID(recordName: extract(record), zoneID: zoneID))
+    }
+  }
+}
